Implement elementary via support to fix #63 tests 2e1 and 2e2.
How it is implemented now:

 - the system accept a via option, and goes is special rendering mode
 - the via-mode returns only the shortest (in time) route
 - the system plans from - via
 - based on the available result it will acquire the arrival time of
   the last ride, its route-idx and trip-offset.
 - the second request bans the last route/trip, this allows to
   search for a vehicle directly after the arrival time of the last leg
 - the second request get the time + walk_slack (+ day)
 - the output is then 'cleaned up' so it appears as one routing request

Downside of the current code: we don't guarantee an optimal (in time)
request. Hence if there was a single itinerary route, that didn't
require any transfers this would make the passenger leave the vehicle.
But, this does allow the use case: "Need to get off the train NOW!"