Alter from_hex function prototype to unsigned char
In qstring.c, the from_hex function was passing a char argument to the isdigit
function. However, according to the C standard (draft N1570), "the argument is
an int, the value of which shall be representable as an unsigned char or shall
equal the value of the macro EOF. If the argument has any other value, the
behavior is undeﬁned". Since char can either be signed or unsigned, casting a
char to an int could lead to negative values (unequal to EOF) being passed to
the isdigit function. This problem was showing up as a cryptic GCC warning about
array indexes, but as it turned out, it was actually a signedness issue instead.