The ANSI raptor.
The target of this branch is to create fully documented, cleaned up
version of our raptor implementation. While we have choosen C99
initially, it seems the portability of this code is still flacky. A
version that is clean in -ansi -pedantic -Wall mode, might be more
portable.

For functions that are not supported unless a feature test is provided,
I have carefully written a substitute which should make porting the code
more easy.