Added --from-date option in gtfs2rrrr script
With this option you can explicitly set what the first valid day in the timetable should be.
This can be used if the GTFS file contains a lot of historical data and you want to use more recent data from it.
The default behaviour is to pick the first 32 days after the first calendar day in the GTFS as it did before.

The script now also prints the last valid day of the timetable.