Fixed handling of absolute times in the following night in realtime updates
An absolute time like 25-3-2015 0:30:00 can't be unambigously converted to an rtime_t.
It can be either part of the previous service day running late into the night.
Or it can be really early in the day (which was the default assumption of epoch_to_rtime() )

Depending on the case the stoptime_t.arrival and stoptime_t.departure should either be
greater or lower than RTIME_ONE_DAY

Wether a time is a day late or just really early can be calculated from the start_date
in the TripDescriptor message.