Re-add sentinel journey_pattern
When a real-time journey_pattern is added it overwrites the sentinel journey_pattern at the end of the tdata->journey_patterns data structure.
This causes code that relies on calculating the amount of journey_pattern_points for the last journey_pattern to fail.
This patch reinstates the sentinel value at the end of the datastructure whenever a new journey_pattern is added