Change normalisation of journey indices back to absolute indices.
The original code made an effort to decrease the bitset required to
store - what the CSA paper refers to as - trip flags. Each boarded
vehicle journey is marked as such. With an onboard flag prevents the
query regarding if it is possible to board the vehicle. This became
relevant upon implementation of precomputed stop wait times.

Precomputed stop wait time substract the wait time at a stop from the
departure time, moving the departure time at the connection before the
arrival time of the previous connection within the same vehicle journey.

The normalisation applied reduced the required trip flags of a one
single working day bitset from 650k to 100k. In broader perspective,
if more days are available the 100k value will move to 650k eventually.
We have benchmarked the change, and the change does not show any
noticeble performance differences.

Absolute indices allow us to additionally mark vehicle journeys
following in the block or in counter clockwise search previous journeys,
to prevent the query regarding the possibily to board, since we are
already on board.